ElasticSearchDestination-v2: native clients moved to client.esnative package
