+++ modules/java-modules/elastic-v2/src/main/java/org/syslog_ng/elasticsearch_v2/messageprocessor/DummyProcessorNative.java
@@ -25,7 +25,7 @@ package org.syslog_ng.elasticsearch_v2.messageprocessor;
 
 import org.elasticsearch.action.index.IndexRequest;
 import org.syslog_ng.elasticsearch_v2.ElasticSearchOptions;
-import org.syslog_ng.elasticsearch_v2.client.ESNativeClient;
+import org.syslog_ng.elasticsearch_v2.client.esnative.ESNativeClient;
 
 public class DummyProcessorNative extends ESNativeMessageProcessor {
