+++ modules/java-modules/elastic-v2/src/main/java/org/syslog_ng/elasticsearch_v2/messageprocessor/ESNativeMessageProcessor.java
@@ -26,7 +26,7 @@ package org.syslog_ng.elasticsearch_v2.messageprocessor;
 import org.apache.log4j.Logger;
 import org.elasticsearch.action.index.IndexRequest;
 import org.syslog_ng.elasticsearch_v2.ElasticSearchOptions;
-import org.syslog_ng.elasticsearch_v2.client.ESNativeClient;
+import org.syslog_ng.elasticsearch_v2.client.esnative.ESNativeClient;
 
 public abstract class ESNativeMessageProcessor implements  ESMessageProcessor {
 	protected ElasticSearchOptions options;
