+++ modules/java-modules/elastic-v2/src/main/java/org/syslog_ng/elasticsearch_v2/messageprocessor/ESSingleNativeMessageProcessor.java
@@ -27,7 +27,7 @@ import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.index.IndexRequest;
 import org.elasticsearch.action.index.IndexResponse;
 import org.syslog_ng.elasticsearch_v2.ElasticSearchOptions;
-import org.syslog_ng.elasticsearch_v2.client.ESNativeClient;
+import org.syslog_ng.elasticsearch_v2.client.esnative.ESNativeClient;
 
 public class ESSingleNativeMessageProcessor extends ESNativeMessageProcessor {
